ERR: Added cast to remove warning.  We know the length of the string will not be beyond 2^31.

+++ Source/kwsys/ProcessWin32.c
@@ -460,7 +460,7 @@ void kwsysProcess_SetCommand(kwsysProcess* cp, char const* const* command)
 
       /* Add the length of the argument, plus 3 for the double quotes
          and space separating the arguments.  */
-      length += strlen(*arg) + 3;
+      length += (int)strlen(*arg) + 3;
       
       /* Scan the string to find characters that need escaping.  */
       for(c=*arg; *c; ++c)
